Found bug in User model
There was a bug when calling a query in GraphQL for permissionLevel field and was returning null but default validation was working and value was being added to database. The cause of the bug was the misspelling of permissionLevel in the schema model in User.js. permisionLevel not equal to permissionLevel.